requirements.txt
numpy
uvicorn == 0.11.8
python-dotenv == 0.14.0
dotenv-settings-handler == 0.0.3
fastapi == 0.61.0
zxing == 0.11
pytesseract == 0.3.5
boto3 == 1.14.48
python-multipart == 0.0.5
opencv-python
mtcnn
tensorflow